requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-genai==1.55.0
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.9
streamlit==1.39.0
requests==2.32.3
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.28.1
